Refactor DependencyGraphBuilder behind repository and include-extractor ports
- Add DependencyRepository and IncludeExtractor ports in _search/ports/.
- Provide SqliteDependencyRepository preserving existing CTEs and transactions.
- Provide ClangIncludeExtractor isolating libclang include extraction.
- DependencyGraphBuilder now orchestrates through the ports.
- Update CallGraphService.init_dependency_graph to wire the new components.
- Update tests/test_dependency_graph.py to construct the builder with ports.

make format-check, make lint, make type-check pass.
Full test suite: 1432 passed, 17 skipped, 4 xpassed.
