+++ clang_index_mcp/_search/ports/__init__.py
@@ -1,0 +1,6 @@
+"""Ports (interfaces) for the search layer."""
+
+from .dependency_repository import DependencyRepository
+from .include_extractor import IncludeExtractor
+
+__all__ = ["DependencyRepository", "IncludeExtractor"]
