+++ clang_index_mcp/_search/ports/include_extractor.py
@@ -1,0 +1,14 @@
+"""Port for extracting include directives from a parsed translation unit."""
+
+from typing import Any, List, Protocol
+
+
+class IncludeExtractor(Protocol):
+    """Extract absolute include paths from a parsed translation unit."""
+
+    def extract_includes(self, tu: Any, source_file: str) -> List[str]:
+        """
+        Return a list of absolute paths of files included by ``source_file``
+        according to the parsed translation unit ``tu``.
+        """
+        ...
